Moving CreateSessionResponse to Responses folder - it's not an entity
