+++ src/BasisTheory.net/Sessions/SessionClient.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 using BasisTheory.net.Common;
 using BasisTheory.net.Common.Entities;
 using BasisTheory.net.Common.Requests;
-using BasisTheory.net.Sessions.Entities;
+using BasisTheory.net.Sessions.Responses;
 using BasisTheory.net.Sessions.Requests;
 using BasisTheory.net.Tokens.Entities;
